MPD connection logic changes and support for multiple entries
Okay, so I decided to move the network discovery connection logic
from the main app to the library as 1) it needed a revamp and
2) I wanted to adhere to MPD's standards. Overall I'm quite happy as to
how it turned out.

Another thing that really bothered me is the support for multiple
entries, for instance like querying songs. This is by no means a must
for the repo, but I also like having a library that works. It's
simplying going down the response and grouping each value that the
reader comes across into the same field name. This is ultimately
the best way to deal with it as MPD's responses are quite basic, and
lack arrays, but using the same index for each value array will come
from the same entry.